Handle modulemd YAML with late mdversion field
Since we need to validate some parts of the format differently
depending on which version we are reading, we need to try to guess
which format is in use in this file. This patch will save this
guess in memory and compare it to the actual value once it is
encountered in the YAML.

The mechanism by which this is accomplished is a terrible hack,
but it is necessary given time-constraints. It will need to be
reworked in the event that we move to a modulemd version 3 at some
point in the future.

Fixes: https://github.com/fedora-modularity/libmodulemd/issues/14

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>