Add Modulemd.Prioritizer
This object can be used to progressively construct a complex setup
of repos with differing module default priorities (such as being
able to override the repodata with a local configuration file).

To use it, create a Modulemd.Prioritizer and call
Prioritizer.add() with the arrays returned by
Modulemd.object_from_file() (or string). Then get the final
result with Prioritizer.resolve().

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>