Fix two incorrect checks for error
The check in modulemd-yaml-parser-defaults.c was both redundant
and wrong. We can only reach that point in the code if the result
value has already been set appropriately, so testing for non-NULL
error is a waste. It also didn't contain a check for whether
error's reference was non-NULL, thus causing a segfault.

While debugging this, I also found a location where the check for
error being initialized had incorrect logic, leading to an
erroneous message in the logs.

Fixes: https://github.com/fedora-modularity/libmodulemd/issues/53

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>