Specification: Allow overriding the default component whitelist
Some packages rename the SRPM and therefore the package build name
during the SRPM build phase.  Module Build Service takes care of package
whitelisting in the tags in manages and therefore it needs to know the
names of the builds when it's setting up the build environment.

This list provides the necessary information to the build system if the
build names do not match the component names.

Signed-off-by: Petr Šabata <contyk@redhat.com>