Work around optimization bug
For some reason, the compiler is optimizing out a memset() that
clears the yaml_event_t at the end of yaml_event_delete(), which
is called implicitly by yaml_emitter_emit(). As a result, the
cleanup at the end of the function ends up performing a
double-free.

This works around the problem by always assuming that the
yaml_event_delete() routine worked properly and just forces the
event.type field to zero, thereby preventing any future calls to
yaml_event_delete() from attempting to free() memory.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>