Deduplicate module streams when merging
This is a fairly hacky way of deduplicating module streams with the
same NSVC that come from different sources (repos). It makes a weak
assumption that if the NSVC is the same, the whole stream is the
same.

A better long-term fix will be to have a full module stream
comparison function, but that will be added to the 2.0 API plans.

Fixes: https://github.com/fedora-modularity/libmodulemd/issues/88

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>