modulemd-yaml-parser: error is not set though it's expected
When parsing the YAML file in _parse_yaml, if an unexpected event is
received, no error is set by YAML_PARSER_PARSE_WITH_ERROR_RETURN (if it
was, it would have already goto "error" label) but
MMD_YAML_ERROR_EVENT_RETURN_RETHROW was used. This means in some cases
`error` could be NULL, though the caller expect it to be set when the
result of the function is false.