Run CI on all releases of Fedora
Sometimes one release is broken by unrelated issues. This will
help us figure out both when libmodulemd has a CI failure and
also other issues in Fedora that can be reported.

This patch also includes some early support for testing on
CentOS 7, but there are several issues specific to that platform
that remain which will be fixed on a subsequent patch and enabled
then.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>