Add locale option to summary and description getters
The ModuleStream.get_summary(), ModuleStream.get_description() and
Profile.get_description() fields need to take an optional locale in
order to return translated strings.

It also drops the properties associated with summary and description
as it is the safest course of action, rather than having users guess
which string they will get back, they must use the getter which will
handle the locale appropriately.

This patch updates the API but does not yet implement the
translation, which will be done later while implementing ModuleIndex
that handles associating the Translation documents with the streams.

Also reflows the gtkdoc comments in modulemd-profile.h.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>