Add python override for ServiceLevel EOL
This override will make set_eol() take a Python datetime.date and
get_eol() return the same.

Reworks the override conditional to just check for API 2+ instead
of presence in the Modulemd namespace, since ServiceLevel existed
in 1.0 as well.

Also adds the overrides to the autopep8 test. This patch includes
a couple autoformating fixes as a result.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>