Upgrade Streams when adding them to the ModuleIndex
This way it will be easier to handle merges, since we can assume they
are all of the same format. Additionally, this patch makes it so that
a mismatch of module names will return a GError rather than aborting.

This patch makes it so that all Streams in an index will be upgraded
to match the highest mdversion of any document that was added. This
is done rather than to upgrade to the latest-known version so that
a libmodulemd update that starts supporting a new version will not
suddenly break consuming applications as long as the modulemd data
in the repositories remains at the older version.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>