Fix memory handling in variant parsing
g_variant_new_*() functions create floating references. This is
intended for exactly the case we're using, where if you pass in a
floating reference into something like a variant array or
dictionary, the container assumes ownership.

The previous implementation was taking full references on the
returned values, which ultimately resulted in excess refs and the
memory not being freed.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>