Handle intent profiles correctly
Previously, if we looked up a set of default profiles by intent
and the intent didn't exist, we were creating it and then
returning NULL for the set of default profiles.

The correct behavior here is to check the intents first and if
they do not exist there, check the common defaults and return
those (or NULL, as appropriate).

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>