API2: Don't fail merges when default streams differ
Instead of failing the merge on a default stream conflict, we will
instead treat it as having no default set. This is safe because if
the module is not yet installed, then this just means its packages
aren't visible and it needs to be selected explicitly.

If some packages from it are already installed, then the module
stream is already set, and thus changing the default won't matter.

Resolves: rhbz#1666871

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>