Add Modulemd.ModuleIndex.dump_to_custom()
In order to support YAML streams from arbitrary sources (such as those
being written to a compressed file through an intermediate library), we
need to support writing through a custom function. This patch enables
consumers of libmodulemd to have indirect access to libyaml's
yaml_parser_set_output() to accomplish this.

Adds a basic test to read through a long string and verify that it
produced an index with the read module.

Related: https://github.com/fedora-modularity/libmodulemd/issues/208

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>