Fix deduplication bug and memory leak
The deduplication had a copy-paste error that was causing streams
that contained an `arch` value to appear not to be duplicates.

Also changes the deduplication test to use the F29 real-world
module metadata which exercises more of this code. In so doing, it
revealed a very subtle memory leak that this patch also addresses.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>