Fix memory issues with XMD handling and overrides
This cleans up the usage of g_variant_ref_sink() and re-syncs the
StreamV1 and StreamV2 to both make a proper deep copy of the XMD
when setting it.

The overrides weren't creating the GVariant arrays and dictionaries
quite correctly. Also, the set_xmd() function had the input marked
as (transfer full), which was incorrect and causing refcounting
issues.

Fixes: https://github.com/fedora-modularity/libmodulemd/issues/290

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>