Check full stream equality when deduplicating
Previously, we assumed that any two streams having the same NSVCA
were identical. Now we can actually test for full equality before
dropping one while deduplicating. Note that this will result in a
hard failure if they aren't identical, because this means that the
enabled repositories are providing conflicting data.

Fixes: https://github.com/fedora-modularity/libmodulemd/issues/188

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>