Add ModuleIndexMerger.resolve_ext()
This will allow us to handle default streams strictly when merging. In most end-user
cases, we want default stream conflicts to simply drop the default stream (this is
fail-safe). But when we are creating a repo with defaults, we want to be stricter and
ensure that we aren't producing a repo with inconsistent data.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>