DOCS: Fix references to common GLib objects
This patch adds support for linking to common GLib objects (fixing
broken links in the installed docs and also including a shell
script to clean up the ones published to the web).

It also adds a new resources section to the docs, mainly for
cross-referencing.

Some changes to the Travis CI and specfile are needed to ensure
that the glib2-doc package is installed at build-time.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>