Add ModuleIndex.update_from_defaults_directory()
This function can be used by pungi and MBS to generate a buildroot for modular and
non-modular builds. It will read all files suffixed with .yaml from the provided
directory and add them to the index. If an overrides_path is provided, those
contents will also be merged in (and preferred in the event of a conflict).

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>