Drop SCANBUILD environment variable
It was intended to make ninja pass the `--status-bugs` argument to
the scan-build tool, but it turns out that it never did this properly
and as of meson 0.52.0, the parsing of that variable changed to
require a full path, breaking the CI builds. This patch makes it
call the standard `ninja scan-build` and I'll work with meson
upstream to add a way to invoke `--status-bugs`.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>