Travis: Run Fedora tests on aarch64
Reorders some of the tests for readability and performance.

The documentation test now runs first so it is updated quickly.

The CentOS tests are now run before the Fedora tests since it is
one of the quickest to finish and will serve as a good canary for
the tests in general.

Also extends the debug test timeout because the extra I/O can take
a lot longer on aarch64, resulting in test flakes.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>