Release.sh: Also bump the version in meson.build
After performing a release, the new version test will immediately
fail because the release has not been bumped yet. This patch bumps
the version prior to pushing the release tag to the upstream
repository.

It also adds a new tool called `semver`, retrieved from
https://github.com/fsaintjacques/semver-tool to handle version
comparisons and version bumps. It is GPLv3, but since it's being
executed as a separate process, it does not make libmodulemd into
a derivative work.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>