Improve RHEL 7 specfile
In order to build for EPEL 7's libmodulemd2 package, we need a few
more changes to the specfile so the same one can be used for RHEL
and Fedora.

This change packages the validator manpage correctly, as well as
adding a %{?v2_suffix} macro that will be set on EPEL 7 to handle
the package, executable and manpage names properly.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>