Always output quoted stream name
Some module streams are named after versions such as "5.30" which
are coerced into numbers (5.3) in other YAML parsers (such as the
default configuration of pyyaml). Since this format is common, we
should avoid outputting ambiguous stream names.

I'm not aware of other fields that would be ambiguous in this way.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>