Change the specification of `ref`
Loosen the restriction on how the build system must treat an
unspecified `ref` in a component. It must always fall back to the
repo's default branch (which may not be 'master' these days), but
if build-system rules exist to pick a different branch, that is
up to the build-system to decide.

Fixes https://pagure.io/modularity/issue/142

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>