Container-based build setup
This adds a new tool ./setup_dev_container.sh to the source root
that will establish a safe environment for experimenting with
libmodulemd without risk to the developer's host system.

It also updates the README.md with some information on how to use
it and fixes the lack of Merlin's credits.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>