:wrench: Modified Scripts to use env for shebang.
The current shebang statements work fine on a Fedora host,
But while trying to setup a development environment on non-Fedora host,
It seems to cause an issue as bash doesn't always live on /bin/bash in some operating systems.

The usage /usr/bin/env bash makes sure, we're pointing towards the environment.

It improves our portability for the scripts for non-fedora-like hosts.

Signed-off-by: nasirhm <nasirhussainm14@gmail.com>