modulemd v2: Relax context up to 13 characters including underscore
cb507415b158b79df44d9b11ca69642c2c546f0b (Add BuildConfig object)
commit broke scratch builds:

$ rhpkg module-scratch-build --file perl-YAML.yaml --buildrequires platform:el8.5.0
Submitting the module build...
Could not execute module_scratch_build: The build failed with:
modulemd-error-quark: Could not validate stream to emit: Non-alphanumeric character in stream context 'a6d43775_1' (1)

The reason is that a scratch build automatically adds an underscore
suffix (e.g.  "_1)" to the context value.

This patch fixes it by allowing the underscore character and
prolonging a limit of 10 characters with 3 additional characters (i.e.
up to 99 scratch builds). We keep v3 context strict as it was (no
underscore, 10 characters) to prevent packagers from specifying
a context in v3 which could not be scratch-built later.
A specification was amended accordingly.

This dichotomy is dictated by the fact that v3 is an input-only
format, while v2 is both input and output format for MBS.

The implemented tests discovered that reading v2 document with
a static context does not perform a validation in contrast to reading
a v3 document. This patch does not change it. It instead marks the
affected tests as TODO and test an explicit validation in addition.

<https://github.com/fedora-modularity/libmodulemd/issues/549>