modulemd-validator: Validate documents using --type option
The tool only supported document types permitted in an index (i.e.
a YUM repository modular metadata). As a result it was impossible to
validate v3 modulemd-packager documents.

This patch adds a --type option to the tool to select and enforce
a specified document type. A default behavior is loading an index.
Example:

modulemd-validator --type modulemd-validator-v3 foo.yaml