Specification: Add /data/demodularized/rpms list to modulemd and modulemd-packager
A motivation is to enable a demodularization of the packages.

Previously if a package was part of a module stream and later the
package was removed, DNF still did not allow installing a same-named
non-modular package.  The reason is that DNF considers all artifacts
from all module version.  Hence simply removing the package from
a newer module version did not exempt the package from a modular DNF
filtering.

There were two options: Either only consider artifacts of the latest
module version, or have an explicit list of removed artifacts in the
module metadata. DNF authors requested the explicit list.

Hence this change amends v2 and v3 modulemd-packager formats and v2
modulemd format with the explicit list. The list does not affect
building a module in any way. It only affects a package manager
whether to mask or not to mask non-modular packages ("modular
filtering" in DNF language).

The list consists of exact RPM package names. I.e. no NEVR strings.
Just package names.

Current DNF plans for interpeting the list:

DNF will use the list for preventing the listed, non-modular packages
from being masked. DNF will only use a list from the latest module
version. This allows module maintainers to get a package back into
a module by removing the package name from the list and rebuilding the
module.

DNF will apply the list to all available modules. Not only to packages
of that stream which delivers the list. That means a module will be
able to demodularize packages of other modules.