Fix a crash when updating an index with an invalid subdocument and a NULL error parameter
If a ModulemdModuleIndex was updated with a YAML document whose "data"
section was invalid, and the updating function was called with a NULL
error argument (e.g. modulemd_module_index_update_from_file(...,
NULL)), the library crashed on a NULL pounter dereference.

This was caused by reusing the error parameter as temporary storage
for subdocument parser errors.

To preserve Glib's semantics of a NULL GError pointer as "do not
report me the detailed errors", this patch makes the
ModulemdModuleIndex loaders robust against NULL error, and it improves
a documentation regarding NULL pointers and memory management in the
ModulemdModuleIndex interface.

Especially, it emphasises that "failures" argument always becomes
allocated and the user must free it even if no failure was
encountered.