Fix clobbering defaults on a modulemd_module_set_defaults() error
Previously, if modulemd_module_set_defaults() experienced an error
because of a mismatching name or unability to upgrade to a requested
defaults version, the module object erroneusly lost its defaults.

This patch fixes it.

It's worth noting that the error path was unreachable from a public
interface. But that could change in future.