Fix modulemd_module_index_upgrade_defaults() to actually use the requested version
Users call modulemd_module_index_upgrade_defaults() to upgrade all the
defaults in an index to a requested defaults version. There was a bug
which caused not upgrading the particular defaults in the index. The
only effect was on metadata of the index.

It hasn't been noticed because defaults support only one version now.

This patch fixes it.