Spec: XML: Move default, obselete, translation into stream level
Also split default to default-profile and default-stream. It makes the
syntax and semantics simpler.

This leverages module-stream nesting. On the other hand with the
split default prolonges notation with factor 1.4.

If this is terrible idea, we should flatten module-stream hierachy
because it otherwise does not have much usage. Having builds in the
hiearchy, but obsoletes and other ones out of it looks odd.