Spec: XML: Constrain default stream intents to unique values
This also introduces defaulting nonexistant attributes (in intents) to
an empty string. It means that <intent/> and <intent name=""/> are
equivalent. This make the serialization less strict. But there is no
way to constrain nonexistant attribute to uniquness in XML Schema.

TODO: Rework other attributes previously forced to require an empty
value to use default attributes.