Spec: XML: Allow empty and missing values for descriptive text
Messages, summaries, descriptions. All free texts can be empty or
missing.

There is probably a distinction in translations where a missing
attribute means unsetting. But practially an empty string as
a traslation does not make much sense. Only when translating an empty
string. We keep empty translations as a translation to allow
translating empty strings. Just for the sake of partity.