Spec: XML: nevra checksums are attributes now
XML Schema is unable to express that an attribute depends on another
attribute while the element has the same name. (I.e. that <checksum
type="sha256" value="abcd"/> is to short value.)

Changing the checksum elements into differently named attributes
solves it. It also makes obvious that a single package cannot have
different checksum values of the same algorithms. (Technically it can,
before and after signing a package. But that's exactly the reason why
people want to remove these checksums from the format.)