Specification: Deprecate intents in modulemd-defaults
System intents which parametrize a default stream and the default
profiles by a system intent (also called a system purpose, system
role, distribution spin) were never implemented in DNF (a package
manager) and never set by Anaconda (a system installer) in Fedora.

In RHEL, system purpose can be set a subscription manager, but it's
purpose over a time become purely statistical and is not used
for enabling repositories or installing packages.

Therefore this feature solely exists in libmodulemd and no modules in
wild use this feature. Thus this commit deprecates it:

It marks "intents" keyword in modulemd-defaults YAML format
deprecated.
It documents that an intent argument of various public will be ignored
in the future.

<https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/RXDP2WMPR3HHBRTQAKPSTRU6KABTJSMA/#JDN2FHCDR6BNPCL4EQKRN3UK54UZOGNN>