Remove an optional dependency on file library (libmagic)
Magic library carries a large database (9 MB). libmodulemd used that
library to detect 3 compresssion formats. Thus depending on that
library was perceived as too expensive.

This patch reimplements the detection of the 3 compression formats
without using magic library.

Because detection of plain text is inherently unreliable (with and
without libmagic), this patch changes modulemd_detect_compression() to
simply return MODULEMD_COMPRESSION_TYPE_NO_COMPRESSION (a plain text)
if the file is not compressed in one of the 3 supported formats. That
also applies to an empty file which was previously reported as
MODULEMD_COMPRESSION_TYPE_UNKNOWN_COMPRESSION (an uknown compression).

This change is considered harmless as a parser error will be reported
later when the file is passed to a YAML parser.

Note that this patch preserves a detection by a file name suffix, which
still takes a precedence over a detection by a file content.