CI: Remove disabling seccomp for a container manager
When glibc-2.34 started to use clone3() syscall, seccomp policy became
violated and glib2 fork+exec functions failed. This discrepancy was
worked around with passing "--security-opt seccomp=unconfined" option
to a container manager.

Now when Microsoft fixed the policy in ubuntu-22.04 images
(ubuntu-20.04 remains broken) and we moved to ubuntu-22.04, the
workaround is not needed. This patch removes it.

<https://github.com/actions/runner-images/issues/3812>