Lessen quoting for git commit IDs
Commit IDs in ref fields starting with a digit were unnecesarily
quoted (ref: "26ca0c0"). This patch augments detecting numbers to
hopefully anything what YAML 1.2.2 considers an integer or a float.
That enables only quoting values which indeed can be interpreted as
numbers and thus are endangered of loosing characters insignificant
for numbers (1.20 -> 1.2).

This adds also an extensive tests to what a number-like string looks
like.