doc: Adapt GLib documentation path to GLib 2.80.1
With GLib 2.80.1 (commit 548ec9f1), the installation paths for the documentation
has changed once again.

Rationale there was:
The `gi-docgen` tool is not designed to be used like that. In
particular, when nesting documentation directories, the generated
`*.devhelp2` files (needed by Devhelp to show the documentation) are
nested one directory level too deep for Devhelp to find them, and hence
are useless, and the documentation doesn’t show up in this common
documentation viewer.