meson: switch version compare logic
meson will automatically suppress the warnings about using
features from later versions in the `if` statement.

This is not the case when the later version features are in
`else`.

Rewrite the conditional to switch to comparing `>=${NEW_VERSION}`
so that meson stops warning the use of newer features than our
`meson_version`.

Signed-off-by: Matt Jolly <kangie@gentoo.org>