CI: Use Packit to build on non-x86_64 architectures
This patch enables Packit to build on AArch64, 32-bit x86, and PPC64LE
platforms on Fedora Rawhide distribution. It also reenables running
tests on AArch64 in the upstream spec file.

This way we should get test coverage for some common non-x86_64
architectures at a reasonable speed.

s390x intentionally skipped because the build waited an hour. That's
unacceptable for CI.